ProductDetailPage divided into smaller components
